airflow/requirements.txt
boto3
configparser
pandas
praw
psycopg2-bin
pyarrow